added weekly and monthly filter
