+++ data.php
@@ -19,11 +19,16 @@ function get_price($name)
 	}
 }
 
-function get_today_promotions(){
+function get_promotions($range){
 
     $conn = connect_db();
 
-	$query = "SELECT * FROM promotions WHERE promotions.expire_date = '".date("Y-m-d")."'";
+	if(empty($range)){
+		$query = "SELECT * FROM promotions WHERE promotions.expire_date = '".date("Y-m-d")."'";
+	}else{
+		$to_date = get_date_range(get_increment($range));
+		$query = "SELECT * FROM promotions WHERE promotions.expire_date BETWEEN '".date("Y-m-d")."' and '".$to_date."'";
+	}
 	
 	if (!$result = $conn->query($query)) {
 		response(500, "Error getting data.", null);
@@ -40,4 +45,21 @@ function get_today_promotions(){
 	}
 
 	response(200,"Ok",$fin_response);    
+}
+
+
+function get_increment($range){	
+	if($range == "week"){
+		$increments = 7-date("N");
+	}elseif($range == "month"){
+		$increments = date("t")-date("j");
+	}
+	return $increments;
+}
+
+function get_date_range($increments){
+	$today = date_create(date("Y-m-d"));
+
+	date_add($today, date_interval_create_from_date_string("".$increments."days"));
+	return $today->format("Y-m-d");
 }
